requirements.txt
certifi==2020.12.5
chardet==3.0.4
idna==2.10
instagrapi==1.3.5
Pillow==8.0.1
praw==7.1.0
prawcore==1.5.0
psycopg2==2.8.6
pydantic==1.7.2
PySocks==1.7.1
requests==2.24.0
si-prefix==1.2.2
six==1.15.0
timeago==1.0.15
update-checker==0.18.0
urllib3==1.25.11
websocket-client==0.57.0
